+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [5.10.0] - 2026-05-04
 ### Added
 - Add schema validation to PURLs (OSIDB-4910)
 - Affect(s) can be automatically created and assigned to Flaw(s) for